+++ mapie/tests/test_multi_label_classification.py
@@ -21,17 +21,20 @@ Params = TypedDict(
     {
         "method": str,
         "bound": Optional[str],
-        "random_state": Optional[int]
+        "random_state": Optional[int],
+        "metric_control": Optional[str]
     }
 )
 
+METHODS = ["crc", "rcps", "ltt"]
+METRICS = ['recall', 'precision']
+
+BOUNDS = ["wsr", "hoeffding", "bernstein"]
 random_state = 42
 
-METHODS = ["crc", "rcps"]
-BOUNDS = ["wsr", "hoeffding", "bernstein"]
-
-WRONG_METHODS = ["rpcs", "rcr", "test", "", 1, 2.5, (1, 2)]
+WRONG_METHODS = ["rpcs", "rcr", "test", "llt"]
 WRONG_BOUNDS = ["wrs", "hoeff", "test", "", 1, 2.5, (1, 2)]
+WRONG_METRICS = ["presicion", "recal", ""]
 
 
 STRATEGIES = {
@@ -39,28 +42,40 @@ STRATEGIES = {
         Params(
             method="crc",
             bound=None,
-            random_state=random_state
+            random_state=random_state,
+            metric_control="recall"
         ),
     ),
     "rcps_wsr": (
         Params(
             method="rcps",
             bound="wsr",
-            random_state=random_state
+            random_state=random_state,
+            metric_control='recall'
         ),
     ),
     "rcps_hoeffding": (
         Params(
             method="rcps",
             bound="hoeffding",
-            random_state=random_state
+            random_state=random_state,
+            metric_control='recall'
         ),
     ),
     "rcps_bernstein": (
         Params(
             method="rcps",
             bound="bernstein",
-            random_state=random_state
+            random_state=random_state,
+            metric_control='recall'
+        ),
+    ),
+    "ltt": (
+        Params(
+            method="ltt",
+            bound=None,
+            random_state=random_state,
+            metric_control='precision'
         ),
     ),
 }
@@ -111,6 +126,17 @@ y_toy_mapie = {
         [True, True, True],
         [True, True, True],
     ],
+    "ltt": [
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+        [False, False, False],
+    ]
 }
 
 
@@ -193,12 +219,26 @@ def test_valid_method() -> None:
     check_is_fitted(mapie_clf, mapie_clf.fit_attributes)
 
 
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+def test_valid_metric_method(strategy: str) -> None:
+    """Test that valid metric raise no errors"""
+    args = STRATEGIES[strategy][0]
+    mapie_clf = MapieMultiLabelClassifier(
+        random_state=random_state,
+        metric_control=args["metric_control"]
+    )
+    mapie_clf.fit(X_toy, y_toy)
+    check_is_fitted(mapie_clf, mapie_clf.fit_attributes)
+
+
 @pytest.mark.parametrize("bound", BOUNDS)
 def test_valid_bound(bound: str) -> None:
     """Test that valid methods raise no errors."""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    mapie_clf = MapieMultiLabelClassifier(
+        random_state=random_state, method="rcps"
+    )
     mapie_clf.fit(X_toy, y_toy)
-    mapie_clf.predict(X_toy, method="rcps", bound=bound, delta=.1)
+    mapie_clf.predict(X_toy, bound=bound, delta=.1)
     check_is_fitted(mapie_clf, mapie_clf.fit_attributes)
 
 
@@ -211,13 +251,14 @@ def test_predict_output_shape(
     """Test predict output shape."""
     args = STRATEGIES[strategy][0]
     mapie_clf = MapieMultiLabelClassifier(
+        method=args["method"],
+        metric_control=args["metric_control"],
         random_state=args["random_state"]
     )
     mapie_clf.fit(X, y)
     y_pred, y_ps = mapie_clf.predict(
         X,
         alpha=alpha,
-        method=args["method"],
         bound=args["bound"],
         delta=.1
     )
@@ -234,13 +275,14 @@ def test_results_for_same_alpha(strategy: str) -> None:
     """
     args = STRATEGIES[strategy][0]
     mapie_clf = MapieMultiLabelClassifier(
+        method=args["method"],
+        metric_control=args["metric_control"],
         random_state=args["random_state"]
     )
     mapie_clf.fit(X, y)
     _, y_ps = mapie_clf.predict(
         X,
         alpha=[0.1, 0.1],
-        method=args["method"],
         bound=args["bound"],
         delta=.1
     )
@@ -257,12 +299,18 @@ def test_results_for_partial_fit(strategy: str) -> None:
     args = STRATEGIES[strategy][0]
     clf = MultiOutputClassifier(LogisticRegression()).fit(X, y)
     mapie_clf = MapieMultiLabelClassifier(
-        estimator=clf, random_state=args["random_state"]
+        estimator=clf,
+        method=args["method"],
+        metric_control=args["metric_control"],
+        random_state=args["random_state"]
     )
     mapie_clf.fit(X, y)
 
     mapie_clf_partial = MapieMultiLabelClassifier(
-        estimator=clf, random_state=args["random_state"]
+        estimator=clf,
+        method=args["method"],
+        metric_control=args["metric_control"],
+        random_state=args["random_state"]
     )
     for i in range(len(X)):
         mapie_clf_partial.partial_fit(
@@ -273,7 +321,6 @@ def test_results_for_partial_fit(strategy: str) -> None:
     y_pred, y_ps = mapie_clf.predict(
         X,
         alpha=[0.1, 0.1],
-        method=args["method"],
         bound=args["bound"],
         delta=.1
     )
@@ -281,7 +328,6 @@ def test_results_for_partial_fit(strategy: str) -> None:
     y_pred_partial, y_ps_partial = mapie_clf_partial.predict(
         X,
         alpha=[0.1, 0.1],
-        method=args["method"],
         bound=args["bound"],
         delta=.1
     )
@@ -299,27 +345,26 @@ def test_results_for_alpha_as_float_and_arraylike(
     """Test that output values do not depend on type of alpha."""
     args = STRATEGIES[strategy][0]
     mapie_clf = MapieMultiLabelClassifier(
+        method=args["method"],
+        metric_control=args["metric_control"],
         random_state=args["random_state"]
     )
     mapie_clf.fit(X, y)
     y_pred_float1, y_ps_float1 = mapie_clf.predict(
         X,
         alpha=alpha[0],
-        method=args["method"],
         bound=args["bound"],
         delta=.9
     )
     y_pred_float2, y_ps_float2 = mapie_clf.predict(
         X,
         alpha=alpha[1],
-        method=args["method"],
         bound=args["bound"],
         delta=.9
     )
     y_pred_array, y_ps_array = mapie_clf.predict(
         X,
         alpha=alpha,
-        method=args["method"],
         bound=args["bound"],
         delta=.9
     )
@@ -332,29 +377,31 @@ def test_results_for_alpha_as_float_and_arraylike(
 @pytest.mark.parametrize("strategy", [*STRATEGIES])
 def test_results_single_and_multi_jobs(strategy: str) -> None:
     """
-    Test that MapieRegressor gives equal predictions
+    Test that MapieClassifier gives equal predictions
     regardless of number of parallel jobs.
     """
     args = STRATEGIES[strategy][0]
     mapie_clf_single = MapieMultiLabelClassifier(
-        n_jobs=1, random_state=args["random_state"]
+        n_jobs=1,
+        metric_control=args["metric_control"],
+        random_state=args["random_state"]
     )
     mapie_clf_multi = MapieMultiLabelClassifier(
-        n_jobs=-1, random_state=args["random_state"]
+        n_jobs=-1,
+        metric_control=args["metric_control"],
+        random_state=args["random_state"]
     )
     mapie_clf_single.fit(X, y)
     mapie_clf_multi.fit(X, y)
     y_pred_single, y_ps_single = mapie_clf_single.predict(
         X,
         alpha=0.2,
-        method=args["method"],
         bound=args["bound"],
         delta=.9
     )
     y_pred_multi, y_ps_multi = mapie_clf_multi.predict(
         X,
         alpha=0.2,
-        method=args["method"],
         bound=args["bound"],
         delta=.9
     )
@@ -378,13 +425,14 @@ def test_valid_prediction(alpha: Any, delta: Any, bound: Any) -> None:
     )
     model.fit(X_toy, y_toy)
     mapie_clf = MapieMultiLabelClassifier(
-        estimator=model, random_state=random_state
+      estimator=model, method="rcps",
+      random_state=random_state
     )
+
     mapie_clf.fit(X_toy, y_toy)
     mapie_clf.predict(
         X_toy,
         alpha=alpha,
-        method="rcps",
         bound=bound,
         delta=delta
     )
@@ -399,19 +447,20 @@ def test_valid_prediction(alpha: Any, delta: Any, bound: Any) -> None:
 @pytest.mark.parametrize(
     "bound", BOUNDS,
 )
-@pytest.mark.parametrize(
-    "method", METHODS,
-)
-def test_array_output_model(method: Any, alpha: Any, delta: Any, bound: Any):
+@pytest.mark.parametrize("strategy", [*STRATEGIES])
+def test_array_output_model(strategy: str, alpha: Any, delta: Any, bound: Any):
+    args = STRATEGIES[strategy][0]
     model = ArrayOutputModel()
     mapie_clf = MapieMultiLabelClassifier(
-        estimator=model, random_state=random_state
+        estimator=model,
+        method=args["method"],
+        metric_control=args["metric_control"],
+        random_state=random_state
     )
     mapie_clf.fit(X_toy, y_toy)
     mapie_clf.predict(
         X_toy,
         alpha=alpha,
-        method=method,
         bound=bound,
         delta=delta
     )
@@ -428,12 +477,14 @@ def test_reinit_new_fit():
 
 
 @pytest.mark.parametrize("method", WRONG_METHODS)
-def test_method_error_in_predict(method: str) -> None:
+def test_method_error_in_fit(method: str) -> None:
     """Test error for wrong method"""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
-    mapie_clf.fit(X_toy, y_toy)
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method=method
+    )
+
     with pytest.raises(ValueError, match=r".*Invalid method.*"):
-        mapie_clf.predict(X_toy, method=method)
+        mapie_clf.fit(X_toy, y_toy)
 
 
 def test_method_error_if_no_label_fit() -> None:
@@ -458,44 +509,92 @@ def test_method_error_if_no_label_partial_fit() -> None:
 
 @pytest.mark.parametrize("bound", WRONG_BOUNDS)
 def test_bound_error_in_predict(bound: str) -> None:
-    """Test error for wrong method"""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    """Test error for wrong bounds"""
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method='rcps'
+    )
+
     mapie_clf.fit(X_toy, y_toy)
     with pytest.raises(ValueError, match=r".*bound must be in.*"):
-        mapie_clf.predict(X_toy, method="rcps", bound=bound, delta=.1)
+        mapie_clf.predict(X_toy, bound=bound, delta=.1)
+
+
+@pytest.mark.parametrize("metric_control", WRONG_METRICS)
+def test_metric_error_in_fit(metric_control: str) -> None:
+    """Test error for wrong metrics"""
+    mapie_clf = MapieMultiLabelClassifier(
+        random_state=random_state,
+        metric_control=metric_control
+    )
+    with pytest.raises(ValueError, match=r".*Invalid metric. *"):
+        mapie_clf.fit(X_toy, y_toy)
 
 
 def test_error_rcps_delta_null() -> None:
     """Test error for RCPS method and delta None"""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method='rcps'
+    )
+
     mapie_clf.fit(X_toy, y_toy)
     with pytest.raises(ValueError, match=r".*delta cannot be ``None``*"):
-        mapie_clf.predict(X_toy, method="rcps")
+        mapie_clf.predict(X_toy)
+
+
+def test_error_ltt_delta_null() -> None:
+    """Test error for LTT method and delta None"""
+    mapie_clf = MapieMultiLabelClassifier(
+        random_state=random_state,
+        metric_control='precision'
+    )
+    mapie_clf.fit(X_toy, y_toy)
+    with pytest.raises(ValueError, match=r".*Invalid delta. *"):
+        mapie_clf.predict(X_toy)
 
 
 @pytest.mark.parametrize("delta", [-1., 0, 1, 4, -3])
 def test_error_delta_wrong_value(delta: Any) -> None:
     """Test error for RCPS method and delta None"""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method='rcps'
+    )
     mapie_clf.fit(X_toy, y_toy)
     with pytest.raises(ValueError, match=r".*delta must be*"):
-        mapie_clf.predict(X_toy, method="rcps", delta=delta)
+        mapie_clf.predict(X_toy, delta=delta)
+
+
+@pytest.mark.parametrize("delta", [-1., 0, 1, 4, -3])
+def test_error_delta_wrong_value_ltt(delta: Any) -> None:
+    """Test error for RCPS method and delta None"""
+    mapie_clf = MapieMultiLabelClassifier(
+        random_state=random_state,
+        metric_control='precision'
+    )
+
+    mapie_clf.fit(X_toy, y_toy)
+    with pytest.raises(ValueError, match=r".*delta must be*"):
+        mapie_clf.predict(X_toy, delta=delta)
 
 
 def test_bound_none_crc() -> None:
     """Test that a warning is raised nound is not None with CRC method."""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method="crc"
+    )
+
     mapie_clf.fit(X_toy, y_toy)
     with pytest.warns(UserWarning, match=r"WARNING: you are using crc*"):
-        mapie_clf.predict(X_toy, method="crc", bound="wsr")
+        mapie_clf.predict(X_toy, bound="wsr")
 
 
 def test_delta_none_crc() -> None:
     """Test that a warning is raised nound is not None with CRC method."""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method="crc"
+    )
     mapie_clf.fit(X_toy, y_toy)
     with pytest.warns(UserWarning, match=r"WARNING: you are using crc*"):
-        mapie_clf.predict(X_toy, method="crc", bound=None, delta=.1)
+        mapie_clf.predict(X_toy, bound=None, delta=.1)
 
 
 def test_warning_estimator_none() -> None:
@@ -508,14 +607,29 @@ def test_warning_estimator_none() -> None:
 @pytest.mark.parametrize("delta", [np.arange(0, 1, 0.01), (.1, .2), [.4, .5]])
 def test_error_delta_wrong_type(delta: Any) -> None:
     """Test error for RCPS method and delta None"""
-    mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
+    mapie_clf = MapieMultiLabelClassifier(
+      random_state=random_state, method="rcps"
+    )
     mapie_clf.fit(X_toy, y_toy)
     with pytest.raises(ValueError, match=r".*delta must be a float*"):
-        mapie_clf.predict(X_toy, method="rcps", delta=delta)
+        mapie_clf.predict(X_toy, delta=delta)
+
+
+@pytest.mark.parametrize("delta", [np.arange(0, 1, 0.01), (.1, .2), [.4, .5]])
+def test_error_delta_wrong_type_ltt(delta: Any) -> None:
+    """Test error for LTT method and delta None"""
+    mapie_clf = MapieMultiLabelClassifier(
+        random_state=random_state,
+        metric_control="precision"
+    )
+
+    mapie_clf.fit(X_toy, y_toy)
+    with pytest.raises(ValueError, match=r".*delta must be a float*"):
+        mapie_clf.predict(X_toy, delta=delta)
 
 
 def test_error_partial_fit_different_size() -> None:
-    """Test error for RCPS method and delta None"""
+    """Test error for partial_fit with different size"""
     clf = MultiOutputClassifier(LogisticRegression()).fit(X_toy, y_toy)
     mapie_clf = MapieMultiLabelClassifier(
         estimator=clf, random_state=random_state
@@ -564,14 +678,18 @@ def test_pipeline_compatibility(strategy: str) -> None:
     )
     pipe.fit(X, y)
     mapie = MapieMultiLabelClassifier(
-        estimator=pipe, random_state=random_state
+        estimator=pipe,
+        method=args["method"],
+        metric_control=args["metric_control"],
+        random_state=random_state
     )
+
     mapie.fit(X, y)
-    mapie.predict(X, method=args["method"], bound=args["bound"], delta=.1)
+    mapie.predict(X, bound=args["bound"], delta=.1)
 
 
 def test_error_no_fit() -> None:
-    """Test error for RCPS method and delta None"""
+    """Test error for no fit"""
     clf = WrongOutputModel()
     mapie_clf = MapieMultiLabelClassifier(
         estimator=clf, random_state=random_state
@@ -584,9 +702,8 @@ def test_error_no_fit() -> None:
 
 
 def test_error_estimator_none_partial() -> None:
-    """Test error for RCPS method and delta None"""
+    """Test error estimator none partial"""
     mapie_clf = MapieMultiLabelClassifier(random_state=random_state)
-
     with pytest.raises(
         ValueError,
         match=r".*Invalid estimator with partial_fit*"
@@ -611,20 +728,83 @@ def test_partial_fit_second_time():
 @pytest.mark.parametrize("strategy", [*STRATEGIES])
 def test_toy_dataset_predictions(strategy: str) -> None:
     """
-    Test that predictions and intervals
-    are similar with two equal values of alpha.
+    Test toy_dataset_predictions.
     """
     args = STRATEGIES[strategy][0]
     clf = MultiOutputClassifier(LogisticRegression()).fit(X_toy, y_toy)
     mapie_clf = MapieMultiLabelClassifier(
-        estimator=clf, random_state=args["random_state"]
+        clf,
+        method=args["method"],
+        metric_control=args["metric_control"],
+        random_state=random_state
     )
     mapie_clf.fit(X_toy, y_toy)
     _, y_ps = mapie_clf.predict(
         X_toy,
         alpha=.2,
-        method=args["method"],
         bound=args["bound"],
         delta=.1
     )
-    np.testing.assert_allclose(y_ps[:, :, 0], y_toy_mapie[strategy])
+    np.testing.assert_allclose(
+            y_ps[:, :, 0],
+            y_toy_mapie[strategy],
+            rtol=1e-6
+        )
+
+
+@pytest.mark.parametrize("method", ["rcps", "crc"])
+def test_error_wrong_method_metric_precision(method: str) -> None:
+    """
+    Test that an error is returned when using a metric
+    with invalid method .
+    """
+    clf = MultiOutputClassifier(LogisticRegression()).fit(X_toy, y_toy)
+    mapie_clf = MapieMultiLabelClassifier(
+        clf,
+        method=method,
+        metric_control="precision"
+    )
+    with pytest.raises(
+        ValueError,
+        match=r".*Invalid method for metric*"
+    ):
+        mapie_clf.fit(X_toy, y_toy)
+
+
+@pytest.mark.parametrize("method", ["ltt"])
+def test_check_metric_control(method: str) -> None:
+    """
+    Test that an error is returned when using a metric
+    with invalid method .
+    """
+    clf = MultiOutputClassifier(LogisticRegression()).fit(X_toy, y_toy)
+    mapie_clf = MapieMultiLabelClassifier(
+        clf,
+        method=method,
+        metric_control="recall"
+    )
+    with pytest.raises(
+        ValueError,
+        match=r".*Invalid method for metric*"
+    ):
+        mapie_clf.fit(X_toy, y_toy)
+
+
+def test_method_none_precision() -> None:
+    clf = MultiOutputClassifier(LogisticRegression()).fit(X_toy, y_toy)
+    mapie_clf = MapieMultiLabelClassifier(
+        clf,
+        metric_control="precision"
+    )
+    mapie_clf.fit(X_toy, y_toy)
+    assert mapie_clf.method == "ltt"
+
+
+def test_method_none_recall() -> None:
+    clf = MultiOutputClassifier(LogisticRegression()).fit(X_toy, y_toy)
+    mapie_clf = MapieMultiLabelClassifier(
+        clf,
+        metric_control="recall"
+    )
+    mapie_clf.fit(X_toy, y_toy)
+    assert mapie_clf.method == "crc"